Add bitlength parameter
Some new 433 MHz RF switches use codes with 32 bit (against the usual 24 bit): this commit allows user to specify the code length and automatically set to 32 bit.
In addition, a verification is implemented that automatically sets the length to 32 bit when a user tries to send a code greater than 24 bits.